+++ src/components/Button/ButtonLink.js
@@ -1,0 +1,19 @@
+import React from 'react'
+import { withRouter } from 'react-router'
+
+import Button from './Button'
+
+const ButtonLink = (props) => {
+  const onClick = e => {
+    e.preventDefault()
+    props.history.push(props.route)
+  }
+
+  return (
+    <Button ver={props.buttonType} addClass={'buttonLink'} onClick={onClick}>
+      {props.children}
+    </Button>
+  )
+}
+
+export default withRouter(ButtonLink)