+++ src/components/Deriv/Deriv.js
@@ -4,7 +4,7 @@ import injectSheet from 'react-jss'
 
 import styles from './styles'
 import Notice from '../Notice/Notice'
-import HelpLink from '../HelpLink/HelpLink'
+import ButtonLink from '../Button/ButtonLink'
 
 import { canonical, siteTitle } from '../../App'
 
@@ -15,9 +15,9 @@ const Deriv = ({classes}) => (
       <meta name='description' content='LanguaDeriv is a tool for modeling word derivation according to arbitrary rules of morphosyntactics.' />
       <link rel='canonical' href={`${canonical}deriv/`} />
     </Helmet>
-    <HelpLink link='/deriv/help'>
+    <ButtonLink route='/deriv/help'>
       Help
-    </HelpLink>
+    </ButtonLink>
     <h2 className='toolTitle'>LanguaDeriv</h2>
     <Notice>This tool is still in planning.</Notice>
   </div>