+++ src/components/Morph/Morph.js
@@ -4,7 +4,7 @@ import injectSheet from 'react-jss'
 
 import styles from './styles'
 import Notice from '../Notice/Notice'
-import HelpLink from '../HelpLink/HelpLink'
+import ButtonLink from '../Button/ButtonLink'
 
 import { canonical, siteTitle } from '../../App'
 
@@ -15,9 +15,9 @@ const Morph = ({classes}) => (
       <meta name='description' content='LanguaMorph is a tool for modeling historical sound change according to arbitrary rules of phonetic change. It can also be used to reverse historical sound change according to those same rules in order to guess at a word’s proto-language roots.' />
       <link rel='canonical' href={`${canonical}morph/`} />
     </Helmet>
-    <HelpLink link='/morph/help'>
+    <ButtonLink route='/morph/help'>
       Help
-    </HelpLink>
+    </ButtonLink>
     <h2 className='toolTitle'>LanguaMorph</h2>
     <Notice>This tool is still in planning.</Notice>
   </div>