+++ src/components/Tree/Tree.js
@@ -4,7 +4,7 @@ import injectSheet from 'react-jss'
 
 import styles from './styles'
 import Notice from '../Notice/Notice'
-import HelpLink from '../HelpLink/HelpLink'
+import ButtonLink from '../Button/ButtonLink'
 
 import { canonical, siteTitle } from '../../App'
 
@@ -15,9 +15,9 @@ const Tree = ({classes}) => (
       <meta name='description' content='LanguaTree is a tool for drawing and labeling syntax trees.' />
       <link rel='canonical' href={`${canonical}tree/`} />
     </Helmet>
-    <HelpLink link='/tree/help'>
+    <ButtonLink route='/tree/help'>
       Help
-    </HelpLink>
+    </ButtonLink>
     <h2 className='toolTitle'>LanguaTree</h2>
     <Notice>This tool is still in planning.</Notice>
   </div>